+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.16.22 (October 18, 2018)
+
+#### :bug: Bug Fix
+
+- `wandb restore` handles the case of not being run from within a git repo.
+- We now open stdout and stderr in raw mode in Python 2 ensuring tools like bpdb work.
+
+#### :nail_care: Enhancement
+
+- PyTorch
+  - Added a new `wandb.hook_torch` method which records the graph and logs gradients & values of pytorch models
+  - `wandb.Image` detects pytorch tensors and uses **torchvision.utils.make_grid** to render the image. 
+
+## 0.16.21 (October 12, 2018)
+
+#### :nail_care: Enhancement
+
+- Catastrophic errors are now reported to Sentry unless WANDB_ERROR_REPORTING is set to false
+- Improved error handling and messaging on startup
+
 ## 0.16.20 (October 5, 2018)
 
 #### :bug: Bug Fix